requirements.txt
asgiref==3.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
djangorestframework==3.12.4
flake8==3.9.0
mccabe==0.6.1
Pillow==8.0.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2020.4
sqlparse==0.4.1
